test_projects/google_gemini_2_0_flash_001_python/requirements.txt

fastapi
uvicorn[standard]
psycopg2-binary
